resolved issue of timer reseting with other state Changes by implementing useContext
